refactor: legal page file name matched component
