+++ packages/curve-ui-kit/src/widgets/Legal/LegalPage.tsx
